pyproject.toml
[project]
name = "rpi-seism"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apprise>=1.9.7",
    "obspy>=1.4.2",
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "pyserial>=3.5",
    "websockets>=16.0",
    "rpi-seism-common",
]

[project.scripts]
rpi-seism = "src.main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
rpi-seism-common = { git = "https://github.com/rpi-seism/rpi-seism-common.git", branch = "main" }
